requirements.txt
pandas
requests
python-dotenv
streamlit
matplotlib
seaborn
statsmodels
scikit-learn
logging
typing 
